+++ src/App.js
@@ -1,27 +1,6 @@
 import logo from './logo.svg';
 import './App.css';
 import Routing from "./routes/index";
-import { getFirestore, collection, getDocs } from 'firebase/firestore';
-import { initializeApp } from "firebase/app";
-import { getAnalytics } from "firebase/analytics";
-
-
-
-const firebaseConfig = {
-  apiKey: "AIzaSyCo0gPK4fEqqPd5WgmhISKfy1P8AgPmf3Y",
-  authDomain: "purduezilla.firebaseapp.com",
-  databaseURL: "https://purduezilla-default-rtdb.firebaseio.com",
-  projectId: "purduezilla",
-  storageBucket: "purduezilla.appspot.com",
-  messagingSenderId: "1043168052788",
-  appId: "1:1043168052788:web:be5443766ed89f9479ac61",
-  measurementId: "G-9G4NCBXCZX"
-};
-
-const fireapp = initializeApp(firebaseConfig);
-const analytics = getAnalytics(fireapp);
-
-const db = getFirestore(fireapp);
 
 
 
@@ -30,8 +9,10 @@ function App() {
     //add in routing in here instead
     <div className="App">
       <Routing />
+      hi lol
     </div>
   );
 }
 
-export default { App, fireapp, analytics, db };
+export default App;
+
